stylin and profilin
Changes made to Allamon Tools:

_variables.scss

        $body-bg: from #FFFFFF to #8e8e8e

need to figure out how to make this a gradient

jquery.bxslider.css

         .bx-wrapper img margin-left:580

this will probably cause an issue for future media problems. Look into
it. Either patch it with media query tags later, or rebuild the slider
with left and right respectively

main.css

        .hero .slider ui li .row .caption .left  : height: 450px

enlarging the caption box to fill the entire section vertically. there
is something going on with a z-index of one of the captions.