worked on the header today
looks like the components section assets > css > scss > components is
generating most of the hassle for the header and such, added some
colors to _variables, and worked on _header.scss and _top-bar.scss
after much searching for ways to override colors for some of the stuff,
also removed the shadow box over the image slider. moved everything up
etc.