Fixin it Up!
From the top down:
For the Bx Slider:
added borders to the little pager controls to make them round, and
brought the size down so they were circles

main.css:
cut down the padding between things, and spaced it out, added shiny
gradients and drop shadows to 3 column labels

main.scss
changed pager stuff too.
added a huge block of styles a couple times because I was not sure how
to fit them in the nest.
h4 and h5 tags for the headers below the slider, p tags are for the
small paragraphs, the mainbord and pad class were added to give borders
and padding to the columns and pics

header.scss:
shrunk the header and made it all fit right

top-bar.scss:
shrunk the last-child margin-right so it could fit with the rest of the
styles

framework.html:

added and removed template lines as required.

almontopnav.html:
removed the active class causing styling confusion

footer.html:
changed the footer to reflect allamon tool service